Gelf logger refactoring
This commit refactors the Gelf logger (previously Gelf formatter),
abstracting the way we pulled the `application`,
`product` and `environment` names, making it easier for the clients of
this library to simply log their needs without needing to constantly messing with these
variables. Additionally, this commit makes mandatory for the users to provide an `application.conf` file in order to
define these variables.

Read configuration directly from load